lab7

+++ lab7_1.cpp
@@ -2,9 +2,9 @@
 using namespace std;
 int main() {
     int a  =  20, b = 30;
-    int c = 4, output;
+    double c = 4, output;
 
-    output = a - float(b/c);
+    output = a - b/c;
     cout << "output: " << output << "\n";
     return 0;
 }
